Move blanket impls into 'digestable' module
The `Digestable` trait is used to constrain blanket impls of `Signer`
and `Verifier` for `digest::Signer` and `digest::Verifier` respectively.

Since that's what it's used for, it makes sense to put the blanket impls
in the same module.
