+++ signature-crate/src/digest/signer.rs
@@ -4,7 +4,7 @@
 //! For use signature algorithms that support an Initialize-Update-Finalize
 //! (IUF) API, such as ECDSA or Ed25519ph.
 
-use super::{Digest, Digestable};
+use super::Digest;
 use crate::{error::Error, signature::Signature};
 
 /// Sign the given prehashed message `Digest` using `Self`.
@@ -15,14 +15,4 @@ where
 {
     /// Sign the given prehashed message `Digest`, returning a signature.
     fn sign_digest(&self, digest: D) -> Result<S, Error>;
-}
-
-impl<S, T> crate::signer::Signer<S> for T
-where
-    S: Digestable + Signature,
-    T: Signer<S::Digest, S>,
-{
-    fn sign(&self, msg: &[u8]) -> Result<S, Error> {
-        self.sign_digest(S::Digest::new().chain(msg))
-    }
 }
